refactor(audit): padronização de imports e estrutura de arquivos
