--- nexus-backend/src/modules/audit/auditModule.ts
@@ -1,6 +1,0 @@
-export * from './entities/auditEntities';
-export * from './enums/auditEnums';
-export * from './dto/auditDto';
-export * from './interceptors/auditInterceptors';
-export * from './audit-log.service';
-export * from './audit.module';
